Split ECS from main schema
- Fix make file to include additional files.
- Modularise schema generation.
